requirements.txt
ccxt
pandas
streamlit
requests
plotly
yfinance
matplotlib
scikit-learn
python-dotenv
